Refactor file upload hook: import formatBytes utility and move it to a separate file
